Fix put Student, Book, and Employee in own file to make more readable
